langfuse-workshop/requirements.txt
flask
langfuse
openai
python-dotenv